pyproject.toml
[project]
name = "cover-letter-cli"
version = "0.1.0"
description = "A CLI tool to generate tailored cover letters from your CV and job descriptions using LLMs."
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "pymupdf",               
    "requests",                
    "readability-lxml",        
    "typer",              
    "python-dotenv",         
    "openai",                 
]

[project.optional-dependencies]
dev = [
    "ruff",                 
]